+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ### Removed
 
+- Dropped Django 4.2 from supported versions.
 - Dropped Django 5.1 from supported versions.
 
 ### Fixed